+++ NEWS.md
@@ -4,6 +4,9 @@
 * `mutate.()`: Can use `stringr::str_glue()` without specifying `.envir`
 * `replace_na.()`: Checks that `replace` arg only uses columns that exist in the data frame
 
+#### Bug fixes
+* `mutate.()`: Can assign to the same column when `.by = character(0)` (#332)
+
 # tidytable 0.6.6
 
 #### Functionality improvements